Removed SideNavPage and replaced with AbstractPage
